Modified the code so that ship structs can be accessed from an array for enhanced modularity
